refactor(services): Modularize services and improve code organization
- Split services into separate modules for better maintainability
- Create dedicated modules for rule, system, backup, cleanup, and template services
- Centralize service constants and improve service registration
- Simplify service setup and unloading process
- Maintain existing service functionality while improving code structure
